backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-decouple>=2.1
django-cors-headers>=4.4
django-filter>=24.2
drf-spectacular>=0.27
django-allauth>=0.63
gunicorn>=22.0
psycopg[binary]>=3.1
redis>=5.0
whitenoise>=6.7
